Updated README.md with run.py changes
